% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binom_test}
\alias{binom_test}
\title{pseudo-vectorized binomial test}
\usage{
binom_test(x, n, p = NULL, ...)
}
\arguments{
\item{x}{x for binom test}

\item{n}{n for binom test}

\item{p}{success probability}

\item{...}{}
}
\value{
a dataframe with statistics from the binom.test.
}
\description{
This function is a wrapper of the \link[stats]{binom.test}
function that can be used with vectors.
}
\examples{

binom_test(x = c(1,2,3),n = c(2,4,6))

}
