% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{compare_signature_sets}
\alias{compare_signature_sets}
\title{Compare 2 sets of signatures}
\usage{
compare_signature_sets(x, y, simplify_set = c("C", "A"), sep = ">")
}
\arguments{
\item{x}{A matrix with signatures as columns and rows with mutation names}

\item{y}{A matrix with signatures as columns and rows with mutation names}

\item{simplify_set}{Reference bases to simplify the mutation names}

\item{sep}{separation between trinucleotide and alternative allele}
}
\value{
Returns a matrix with pair-wise cosine similarity between signatures
}
\description{
Each signature of the 2 matrices is compared against each of the
signatures of the other matrix using cosine similarity.
}
