% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indels.R
\name{compute_mh_length}
\alias{compute_mh_length}
\title{Compute micro-homology length}
\usage{
compute_mh_length(vec1, vec2, direction = "+")
}
\arguments{
\item{vec1}{a string sequence vector}

\item{vec2}{a string sequence vector}

\item{direction}{in which direction to check for homologies (either + ot -)}
}
\value{
A integer value with the length of the microhomologies
}
\description{
Given 2 seuence vectors it computes the lenght of the micro-homology
if existent.
}
\details{
Should be maybe also able to get a DNAstringSet
}
\examples{
seq1 = c("A","A","T")
seq2 = c("A","A","C")
compute_mh_length(vec1 = seq1,vec2 = seq2,direction = "-")
}
