% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_rAPOBEC}
\alias{compute_rAPOBEC}
\title{Compute apobec enrichment locally for a set of mutations}
\usage{
compute_rAPOBEC(vr, genome = genome_selector(), wl = 1e+06)
}
\arguments{
\item{vr}{a VRanges object with a set of mutations}

\item{genome}{a BSgenome object}

\item{wl}{the window lenght to explore}
}
\value{
A single integer indicating the ratio of rates. Higher values indicate
more apobec3a/b enrichment.
}
\description{
Implemented from the methods in Seplyarskiy 2016 Genome Research.
}
\details{
It computes the number of APOBEC mutations defined a TCW>K and it compares
these numbers to nonAPOBEC mutations in the VCW>K context surrounding the
given set of mutations. The mutations are extended with a wl described in
the arguments. The ratios are corrected by context occurence.
}
