% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muts.R
\name{correct_MSM_P}
\alias{correct_MSM_P}
\title{Correct MSM per total positives}
\usage{
correct_MSM_P(msm, P, ...)
}
\arguments{
\item{msm}{MSM matrix from compute_MSM_fast}

\item{P}{vector with the positives per row}

\item{...}{}
}
\value{

}
\description{
Correct MSM per total positives
}
