% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muts.R
\name{count_MS}
\alias{count_MS}
\title{Count mutaiton subtypes}
\usage{
count_MS(ms)
}
\arguments{
\item{ms}{A mutation subtype string vector}
}
\value{
a integer vector with frequency counts
}
\description{
Given a mutation subtype string vector, automatically indentify the
parameters and generate a interger vector with counts.
}
\examples{

count_MS(c("TCA>T","CAC>T"))
count_MS(c("TCA_T","CAC_T"))
}
