% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indels.R
\name{detect_microhomology}
\alias{detect_microhomology}
\title{Detect microhomology}
\usage{
detect_microhomology(seq_vector)
}
\arguments{
\item{seq_vector}{}
}
\value{
A data frame, row per element in the input vector with the detected
micro-homology.
}
\description{
From a given sequence vector detect microhomologies and return
the lenght of the positive and negative micro-homologies.
}
\examples{
seqvec = c("CTACTCGTTCGA","TTTCCTACTATTTCCGGGTAGGGGGGGGTA")
res = detect_microhomology(seqvec)
}
