% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dna_codes}
\alias{dna_codes}
\title{IUPAC DNA codes}
\format{
A list with all IUPAC DNA codes and its meanings
}
\source{
\url{https://www.bioinformatics.org/sms/iupac.html}
}
\usage{
dna_codes
}
\description{
A list of each DNA letter and its representative set in ACTG letters.
}
\keyword{datasets}
