% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{download_signature_set}
\alias{download_signature_set}
\title{Download cosmic signature set}
\usage{
download_signature_set(
  version = c("v2", "v3"),
  synapse_email = Sys.getenv("SYNAPSE_EMAIL"),
  synapse_apiKey = Sys.getenv("SYNAPSE_PAT")
)
}
\arguments{
\item{version}{Either v2 or v3 indicating the version of the signatures}

\item{synapse_email}{Better use SYNAPSE_EMAIL environtment variable}

\item{synapse_apiKey}{Better use SYNAPSE_PAT environtment variable}
}
\value{
a matrix with selected set of mutational signatures from cosmic
}
\description{
It downloads the matrix with mutation signatures profiles
from cosmic. Version 2 and 3 are available.
}
\details{
For v3, signatures are stored in \url{synapse}{https://www.synapse.org/}.
Therefore, to download such signatures you should include an apiKey and
your email. To obtain the apiKey, log in, go to profile, settings and
API key
}
\references{
Go to the following link for more information: \url{https://cancer.sanger.ac.uk/cosmic/signatures}
}
