% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muts.R
\name{generate_mut_types}
\alias{generate_mut_types}
\title{Generate Mutation Subtypes}
\usage{
generate_mut_types(k, sep = ">", simplify_set = c("C", "T"))
}
\arguments{
\item{k}{The number of upstream and downstream bases the motif is generated}

\item{sep}{The string that separates the motif to the alternative allele.}

\item{simplify_set}{The reference base set to base the mutation subtypes}
}
\value{
A complete mutation set
}
\description{
Generate all the possible mutation subtypes in a mutation set.
}
\examples{
trincl = generate_mut_types(k = 1)
length(trincl) # should be 96

pentancl = generate_mut_types(k = 2)
length(pentancl) # should be 1536

}
