% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{genome_selector}
\alias{genome_selector}
\title{Genome Selector}
\usage{
genome_selector(
  alias = "Hsapiens.UCSC.hg19",
  install = FALSE,
  available = FALSE,
  script_mode = FALSE,
  installed = FALSE
)
}
\arguments{
\item{alias}{string containing the alias of the genome (ex: Hsapiens.NCBI.GRCh38)}

\item{install}{boolean if genome has to be installed in the system if required.}

\item{available}{List available genomes in bioConductor}

\item{script_mode}{Don't toggle the interactive mode.}

\item{installed}{List installed genomes in the local machine}
}
\value{
A BSgenome object.
}
\description{
Wrapper function for a BSgenome usage. It can handle interactive downloads, 
listing of available and installed reference sequences.

Main function is to retrieve a genome object without loading the 
whole package, can be done through aliases or interactive selection.
}
\examples{

\dontrun{
# list installed genomes in the system
# choose one to save it as an object
genome = genome_selector(installed = T)

# how to install a genome package
# then press the desired genome package and press 1 to install.
# after that it will be saved as an object
genome = genome_selector(available = T)


# select the desired genome package
genome = genome_selector(alias = "Hsapiens.1000genomes.hs37d5")
}

}
