% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muts.R
\name{get_MS_VR}
\alias{get_MS_VR}
\title{get mutation subtypes from VR}
\usage{
get_MS_VR(
  x,
  sep = ">",
  k = 1,
  genome,
  simplify_set = c("C", "A"),
  keep_strand = FALSE
)
}
\arguments{
\item{x}{a VRanges object}

\item{sep}{a separator string to divide context and alternative allele}

\item{k}{number of nucleotides to expand the context in one direction}

\item{genome}{a BSgenome object}

\item{simplify_set}{a set of DNA bases which to simplify the mutation strings}

\item{keep_strand}{true if no simplification is needed}
}
\value{
A string vector with mutation strings corresponding to each position in the
VRanges object. Should have the same length as the original inputed
object.
}
\description{
get mutation subtypes from VR
}
\examples{

library(VariantAnnotation)
library(helperMut)
genome = genome_selector("Hsapiens.1000genomes.hs37d5")
vr = VRanges(seqnames = c(1,2,3),
            ranges = IRanges(start = c(1e6,1e6,1e6),width = 1),
            ref = "N", alt = c("A","T","A"))
get_MS_VR(vr,genome = genome)
get_MS_VR(vr,genome = genome,sep = "_")
get_MS_VR(vr,genome = genome,sep = "_",keep_strand = TRUE)
get_MS_VR(vr,genome = genome,simplify_set = c("C","T"))

}
