% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{get_region_chunks}
\alias{get_region_chunks}
\title{Get chunks from regions}
\usage{
get_region_chunks(gr, wl = 10000, unlist = TRUE)
}
\arguments{
\item{gr}{a GRanges object or a BSgenome object}

\item{wl}{the window length of the resulting chunks}

\item{unlist}{if result should be returned as a single GR, default true}
}
\value{
A GRanges object chunked
}
\description{
From a genome file or a GR object generate a chunked version
}
