% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muts.R
\name{identify_mut_aestetics}
\alias{identify_mut_aestetics}
\title{Detect mutation subtype params}
\usage{
identify_mut_aestetics(ms, force = FALSE)
}
\arguments{
\item{ms}{A mutation subtype version (any k)}

\item{force}{Avoid error if non-standard mutations are used as a central position, such as TNA>C}
}
\value{
A list with the extracted parameters
}
\description{
From a single string of mutations detect which parameters, such as k and
sep are being used.
}
\details{
It also identifies the reference set of bases used as central position.s
}
\examples{

helperMut::identify_mut_aestetics(c("TCA>T","CCA>T","TTT>A"))

}
