% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard Index}
\usage{
jaccard(x, y)
}
\arguments{
\item{x}{a character ector}

\item{y}{a second character vector}
}
\value{
the Jaccard index
}
\description{
Computes the jaccard index between two string vectors.
}
\examples{

set1 = c("a","b","c")
set2 = c("b","d","t")
jaccard(set1,set2)

}
