% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muts.R
\name{make_set}
\alias{make_set}
\title{Make mutation set}
\usage{
make_set(
  x,
  simplify = TRUE,
  simplify_set = c("C", "A"),
  sep = ">",
  strict = TRUE
)
}
\arguments{
\item{x}{a simplified version of a mutation set in UPAC format or a vector of non-overlaping sets.}

\item{simplify}{Boolean to simplify or not the result}

\item{simplify_set}{Simpliy reference set if simplify is true.}

\item{sep}{string to separate the context and the alternative allele}

\item{strict}{removes impossible mutation types from the output}
}
\value{
a mutation set with all the possible mutation types encoded.
}
\description{
Make mutation set
}
\examples{

make_set("WAW>T")

}
