% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muts.R
\name{mut2ms6}
\alias{mut2ms6}
\title{Muts to MS6}
\usage{
mut2ms6(mutation_vec, sep = ">")
}
\arguments{
\item{mutation_vec}{a string vector with mutation strings}

\item{sep}{a separator used in the mutation string}
}
\value{
Returns a vector of same length with the mutation types at k=0
}
\description{
Transforms any mutation level (any k) to ms6 (a.k.a k=0)
}
\examples{
mut2ms6("TCA>T") # returns C>T

}
