% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muts.R
\name{mutation_enrichment_general}
\alias{mutation_enrichment_general}
\title{Compute test for mutation enrichment in a region}
\usage{
mutation_enrichment_general(
  vr,
  gr,
  genome = genome_selector(),
  genome_mask = NULL,
  ...
)
}
\arguments{
\item{vr}{a VRanges object with mutations}

\item{gr}{a target region to test for enrichment, a GRanges object}

\item{genome}{a BSgenome object}

\item{genome_mask}{a regions mask that contains the target regions to test.}

\item{...}{arguments for the fisher test}
}
\value{
a tidy dataframe from broom::tidy.
}
\description{
This functions computes a fisher test to check for an enrichment
or depletion based on the size of the region tested and the genome
or genomic mask.
}
\details{
It does not control for sequence composition, just for size.

Warning: target regions will be intercepted with mask if not NULL
}
\examples{

library(VariantAnnotation)
genome = genome_selector("Hsapiens.UCSC.hg19")
base_pos = 6e4
vr_target = VRanges(seqnames="chr1",
                    ranges=IRanges(c(base_pos + 6, base_pos + 16),
                                   width = 1),
                    ref = "C",alt = "A")
gr_target <- GRanges(seqnames="chr1",
                     ranges=IRanges(base_pos + 3,base_pos +  10))
gr_mask <- GRanges(seqnames=c("chr1", "chr1"),
                   ranges=IRanges(c(base_pos + 4,base_pos + 15),
                                  width = 10))

mutation_enrichment_general(vr = vr_target,
                            gr = gr_target,
                            genome = genome,
                            genome_mask = gr_mask)
}
