% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{obtain_genomic_feature}
\alias{obtain_genomic_feature}
\title{Obtain genomic regions from gene lists}
\usage{
obtain_genomic_feature(
  feat,
  annotation_pkg,
  transcript_db,
  identifiers,
  indentifier_type,
  transcript_db_id_type = "ENTREZID",
  transcript_db_columns = NULL,
  ...
)
}
\arguments{
\item{feat}{the genomic feature to extract from (such as genes, transcripts...) see ?GenomicFeatures::transcripts}

\item{annotation_pkg}{An OrgDb object}

\item{transcript_db}{A TxDb object}

\item{identifiers}{the genes that we want to retrieve}

\item{indentifier_type}{the input identifier column}

\item{transcript_db_id_type}{the identifier type for the txdb object}

\item{transcript_db_columns}{extra columns to retrieve from the txdb object}

\item{...}{extra options for feat function}
}
\description{
Wrapper for a workflow that obtains a given genomic feature
from some gene identifier.
}
\details{
Basically it comprises 2 steps, first, transform input
gene ids (such as SYMBOL) to the gene id type for the
txdb object. Second, it extracts the given feature filtering
for those genes inputed.
}
\examples{

\dontrun{
# retrieving from ensembl
human = org.Hs.eg.db::org.Hs.eg.db
txdb_ensembl = makeTxDbFromEnsembl(organism = "homo_sapiens")
obtain_genomic_feature(feat = "transcripts",
                       annotation_pkg = human,
                       transcript_db = txdb_ensembl,
                       identifiers = c("BRCA1","BRCA2"),
                       indentifier_type = "SYMBOL",
                       transcript_db_id_type = "ENSEMBL",
                       transcript_db_columns = "TXNAME")
}

# retrieving from UCSC (faster)

human = org.Hs.eg.db::org.Hs.eg.db
txdb = TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene
obtain_genomic_feature(feat = "transcripts",
                       annotation_pkg = human,
                       transcript_db = txdb,
                       identifiers = c("BRCA1","BRCA2"),
                       indentifier_type = "SYMBOL",
                       transcript_db_columns = "TXNAME")

}
