% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{order_ms96_cosmicSignatures}
\alias{order_ms96_cosmicSignatures}
\title{Ordered mutation types based on Cosmic signatures website}
\format{
A vector with 96 mutation types
}
\source{
\url{https://cancer.sanger.ac.uk/signatures_v2/Signature-2.png}
}
\usage{
order_ms96_cosmicSignatures
}
\description{
A vector containing the ordered mutation types at k=1
from the sanger cosmic website
}
\keyword{datasets}
