% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{order_ms96_supekCell2017}
\alias{order_ms96_supekCell2017}
\title{Ordered mutation types based on Supek,Lehner Cell 2017}
\format{A vector with 96 mutation types}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S0092867417307742}
}
\usage{
order_ms96_supekCell2017
}
\description{
A vector containing the ordered mutation types at k=1
coming from Supek,Lehner Cell 2017
}
\keyword{datasets}
