% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{shufle_regions}
\alias{shufle_regions}
\title{Shufle BED}
\usage{
shufle_regions(regions, mask)
}
\arguments{
\item{regions}{regions object to be shufled (Granges)}

\item{mask}{mask regions where to shufle (Granges)}
}
\value{
The function returns a GR object with the regions located in new random 
positions. The index of the output should match index of the input.

Chromosome / seqnames are not maintained.

Resulting regions may overlap
}
\description{
Similar to \href{https://bedtools.readthedocs.io/en/latest/content/tools/shuffle.html}{shufleBed} from bedtools.
}
\examples{

}
