% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muts.R
\name{ms_reverse_complement}
\alias{ms_reverse_complement}
\alias{str_reverse_complement}
\title{Reverse complement of a string}
\usage{
ms_reverse_complement(ms)

str_reverse_complement(str)
}
\arguments{
\item{ms}{a mutation character vector (assumes 1 separation character)}

\item{str}{a sequence character vector}
}
\value{
A string with the reverseComplement of the input
}
\description{
It returns the reverse complement of a context or mutation string.
}
\details{
Internally it transforms the string to DNAstringSet from Biostrings.
}
\section{Functions}{
\itemize{
\item \code{ms_reverse_complement()}: Rc of a mutation string

}}
\examples{

str_reverse_complement(c("CCC","TTT"))

ms_reverse_complement(c("CCC>T","TTT>A"))

}
