% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tr_colors}
\alias{tr_colors}
\title{Colors used to plot mutational signatures}
\format{A names vector with 6 hex codes.}
\source{
\url{https://cancer.sanger.ac.uk/signatures_v2/Signature-2.png}
}
\usage{
tr_colors
}
\description{
A vector containing the HEX codes for each mutation types at k=0.
}
\details{
It can directly be used in scale_manual from ggplot2
}
\keyword{datasets}
