% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tr_colors_CA}
\alias{tr_colors_CA}
\title{Color scheme for CA mutations}
\format{
A vector or list of color codes
}
\usage{
data(tr_colors_CA)
}
\description{
A dataset containing color codes for CA mutations.
}
\keyword{datasets}
