% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tr_colors_CT}
\alias{tr_colors_CT}
\title{Color scheme for CT mutations}
\format{
A vector or list of color codes
}
\usage{
data(tr_colors_CT)
}
\description{
A dataset containing color codes for CT mutations.
}
\keyword{datasets}
