% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vr_example}
\alias{vr_example}
\title{A small VR object to use in examples.}
\format{
VRanges object
}
\usage{
vr_example
}
\description{
A small VRanges object to use in the examples of the documentation
}
\keyword{datasets}
