% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muts.R
\name{extend}
\alias{extend}
\title{GRanges extension}
\usage{
extend(x, upstream, downstream)
}
\arguments{
\item{x}{A GRanges or VRanges object}

\item{upstream}{Number of nucleotides to extend upstream-wise}

\item{downstream}{Number of nucleotides to extend downstream-wise}
}
\value{
A GRanges or VRanges object
}
\description{
The function extend all the ranges in a \link[GenomicRanges]{GRanges} object by the number
of nucleotides defined in the upstream or downstream of each range.

It also works for \link[VariantAnnotation]{VRanges} objects.

extracted from \url{https://support.bioconductor.org/p/78652/}
}
