% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_k_freq}
\alias{get_k_freq}
\title{Get k-mer frequency in neighbouring regions}
\usage{
get_k_freq(vr, k, wl, genome = genome_selector(), ...)
}
\arguments{
\item{vr}{a VRanges object with SNVs}

\item{k}{the extension of mutation context in one direction}

\item{wl}{the window length from which each mutation would be expanded}

\item{genome}{a BSgenome object}

\item{...}{arguments for oligonucleotideFrequency}
}
\value{

}
\description{
From mutations in a VRanges object, obtain the k-mer oligonucleotide
frequencies at a given window length.
}
