% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr_example}
\alias{gr_example}
\title{A small GR object to use in examples.}
\format{GRanges object}
\usage{
gr_example
}
\description{
A small GRanges object to use in the examples of the documentation
}
\keyword{datasets}
