% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indels.R
\name{indels_classifier}
\alias{indels_classifier}
\title{Indel Classifier}
\usage{
indels_classifier(
  vr,
  maxRep = 5,
  genome = genome_selector(),
  check_format = TRUE,
  groupVar = "group"
)
}
\arguments{
\item{vr}{A VRanges object with left-aligned indels}

\item{maxRep}{The number of max repretitions for the indels}

\item{genome}{A BSgenome object}

\item{check_format}{Should we check the indel input format?}

\item{groupVar}{Used to determine rows in the final matrix}
}
\value{
A dataframe with rows as grouped elements and columns as features
}
\description{
UNDER DEVELOPMENT!
}
\details{
Given a set of left-aligned indels the function classifies the mutations
according to he indels rules defined by PCAWG (?) to generate indel features.

See here for more details.

Input format needs to be first letter in REF and ALT is shared. Then this
letter corresponds to the reference sequence in genome. This should be the
format obtained from
\url{bcftools norm}{http://samtools.github.io/bcftools/bcftools.html#norm}
}
