% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muts.R
\name{simplify_muts}
\alias{simplify_muts}
\alias{simplify_ctx}
\title{Simplify mutations}
\usage{
simplify_muts(muts, simplify_set = c("C", "A"), sep = ">")

simplify_ctx(ctx, simplify_set = c("C", "A"))
}
\arguments{
\item{muts}{A character vector containing mutations}

\item{simplify_set}{The reference base set. (must be 2)}

\item{sep}{The string that separates context and alternative}
}
\value{
A set of simplified mutations
}
\description{
Given a set of mutations it removes the strand information and simplifies them.
It also translate a mutation set to an alternative refernce set (for instance, C T based to C A based.)
}
\section{Functions}{
\itemize{
\item \code{simplify_ctx}: Simplify contexts
}}

\examples{

mutset = c("ACT>T","TTA>C")
simplify_muts(mutset)

ctxset = c("ACT","TTA")
simplify_ctx(ctxset)

}
